feat(lobby-notification): Migrate lobby notifications to the standard.
